fix: move main function to cli file
